+++ tests/unit/database_tests.py
@@ -202,6 +202,11 @@ class DatabaseTests(UnitTestDbBase):
             # No issue should arise if attempting to create existing database
             db_2 = db.create()
             self.assertEqual(db, db_2)
+            # If we use throw_on_exists=True, it will raise a
+            # CloudantDatabaseException if the database already exists.
+            with self.assertRaises(CloudantDatabaseException) as cm:
+                db.create(throw_on_exists=True)
+            self.assertEqual(cm.exception.status_code, 412)
         except Exception as err:
             self.fail('Exception {0} was raised.'.format(str(err)))
         finally:
@@ -690,7 +695,7 @@ class DatabaseTests(UnitTestDbBase):
         )
         # Test differences
         self.assertEqual(
-            self.db.revisions_diff('julia006', *revs), 
+            self.db.revisions_diff('julia006', *revs),
             {'julia006': {'missing': revs, 'possible_ancestors': [doc['_rev']]}}
         )
         # Test no differences
@@ -1150,7 +1155,7 @@ class CloudantDatabaseTests(UnitTestDbBase):
                  'lists': {},
                  'shows': {},
                  'language': 'query',
-                 'views': {index.name: {'map': {'fields': {'name': 'asc', 
+                 'views': {index.name: {'map': {'fields': {'name': 'asc',
                                                            'age': 'asc'}},
                                         'reduce': '_count',
                                         'options': {'def': {'fields': ['name',
@@ -1241,7 +1246,7 @@ class CloudantDatabaseTests(UnitTestDbBase):
                  'lists': {},
                  'shows': {},
                  'views': {'json-index-001': {
-                                'map': {'fields': {'name': 'asc', 
+                                'map': {'fields': {'name': 'asc',
                                                    'age': 'asc'}},
                                         'reduce': '_count',
                                         'options': {'def': {'fields': ['name',
@@ -1258,7 +1263,7 @@ class CloudantDatabaseTests(UnitTestDbBase):
                                    'default': 'keyword',
                                    'fields': {'$default': 'standard'}}}}}
             )
-    
+
     def test_create_query_index_failure(self):
         """
         Tests that a type of something other than 'json' or 'text' will cause